refactored the dir structure
