+++ frontend/src/App.js
@@ -1,10 +1,17 @@
 import React from 'react';
+import { Route, BrowserRouter as Router } from 'react-router-dom';
+
 import withFirebaseAuth from 'react-with-firebase-auth';
 import firebase from 'firebase';
 import 'firebase/auth';
 import firebaseConfig from './firebaseConfig';
-import Homepage from './components/homepage';
-// import SignUp from './components/signUp'
+
+import Homepage from './landing/homepage';
+import SignUp from './landing/signUp';
+import LoggedIn from './authed/main';
+import LoginComponent from './landing/login';
+import LogoutComponent from './landing/logout';
+import CreateComponent from './landing/create';
 
 const firebaseApp = firebase.initializeApp(firebaseConfig);
 const firebaseAppAuth = firebaseApp.auth();
@@ -14,9 +21,16 @@ const providers = {
 
 function App() {
   return (
-    <div>
-      <Homepage />
-    </div>
+    <Router>
+      <div>
+        <Route exact path="/" component={Homepage} />
+        <Route path="/signup" component={SignUp} />
+        <Route path="/auth" component={LoggedIn} />
+        <Route path="/login" component={LoginComponent} />
+        <Route path="/logout" component={LogoutComponent} />
+        <Route path="/create" component={CreateComponent} />
+      </div>
+    </Router>
   );
 }
