+++ frontend/src/authed/dashboard/handlers/InputHooks.js
@@ -1,4 +1,4 @@
-import { useState } from 'react'; // removed React from here just to get rid of warnings
+import { useState } from 'react';
 
 const useSignUpForm = (initialValues, callback) => {
   const [inputs, setInputs] = useState(initialValues);
