+++ frontend/src/authed/profile.js
@@ -1,11 +1,9 @@
 import React from 'react';
 
-import ResumeDashboard from './resume/dashboard';
-
 const Profile = () => {
   return (
     <div>
-      <ResumeDashboard />
+      Profile
     </div>
   );
 };
