+++ frontend/src/index.js
@@ -1,23 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import { Route, BrowserRouter as Router } from 'react-router-dom';
-import SignUp from './components/signUp';
-import Nav from './nav/navigation';
-import LoginComponent from './components/login';
-import LogoutComponent from './components/logout';
-import CreateComponent from './components/create';
+import * as serviceWorker from './serviceWorker';
 
-const routing = (
-  <Router>
-    <div>
-      <Route exact path="/" component={App} />
-      <Route path="/signup" component={SignUp} />
-      <Route path="/nav" component={Nav} />
-      <Route path="/login" component={LoginComponent} />
-      <Route path="/logout" component={LogoutComponent} />
-      <Route path="/create" component={CreateComponent} />
-    </div>
-  </Router>
-);
-ReactDOM.render(routing, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
